ref(irc): Split client internals into modules
irc.ts had turned into the place where options, public types, byte helpers, isupport parsing, and the client class all lived together.

Pull the shared pieces into internal modules and replace the inferred IrcOptions type with an explicit documented interface so the public API reads a lot better in editors.

This keeps behavior the same, but makes the next round of client refactoring a lot less gross.
